+++ src/main/java/me/pajic/affogato_core/CompatFlags.java
@@ -7,5 +7,6 @@ public class CompatFlags {
     public static final boolean ITEMSWAPPER_LOADED = FabricLoader.getInstance().isModLoaded("itemswapper");
     public static final boolean SERENE_WILD_LOADED = FabricLoader.getInstance().isModLoaded("serenewild");
     public static final boolean WILDER_WILD_LOADED = FabricLoader.getInstance().isModLoaded("wilderwild");
+    public static final boolean NULLSCAPE_LOADED = FabricLoader.getInstance().isModLoaded("nullscape");
     public static final boolean MASTERCUTTER_LOADED = FabricLoader.getInstance().isModLoaded("mr_mastercutter");
 }